update tests for request util
